src/app/core/guard/guest/guest.guard.ts
import { Injectable } from '@angular/core';
import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot, UrlTree } from '@angular/router';
import { Observable } from 'rxjs';

@Injectable({
  providedIn: 'root'
})
export class GuestGuard implements CanActivate {
  constructor(
    private readonly router: Router,
  ){}
  canActivate(): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
      let token = window.localStorage.getItem('token');
      if(token === null) {
        return true;
      } else {
        this.router.navigate(['/admin']);
        return false;
      }
  }
  
}
